Update patch set 1
Patch Set 1:

I'm curious as to how much of a drawback that exactly is for this particular mode. Are we talking 10x the power usage? 2x the usage? Personally I don't believe it would be that much noticeable, given that this mode is not meant to be used all the time.

Patch-set: 1
Reviewer: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>